Removed dependencies from repository

+++ Sudoku/Line.cpp
@@ -7,7 +7,7 @@ Line::Line(const sf::Vector2f& point1, const sf::Vector2f& point2) :
 	sf::Vector2f unitDirection = direction / std::sqrt(direction.x * direction.x + direction.y * direction.y);
 	sf::Vector2f unitPerpendicular(-unitDirection.y, unitDirection.x);
 
-	sf::Vector2f offset = (m_thickness / 2.f) * (unitPerpendicular);
+	sf::Vector2f offset = (static_cast<float>(m_thickness) / 2.f) * (unitPerpendicular);
 
 	m_vertices[0].position = point1 + offset;
 	m_vertices[1].position = point2 + offset;
